pyproject.toml
[tool.poetry]
name = "odtp-shacl-maker"
version = "0.0.1"
description = "SHACL generator for odtp compatible components"
authors = ["caviri <carlosvivarrios@gmail.com>"]
license = "AGPL 3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
typer = "0.12.3"
pandas = "^2.2.2"
rdflib = "^7.0.0"
yamlreader = "^3.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
